+++ src/viewpoint-measures/HeuristicMeasure.cpp
@@ -12,14 +12,14 @@ HeuristicMeasure::HeuristicMeasure(const QString& pName): Measure(pName, true)
 void HeuristicMeasure::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
 {
     const auto projectedAreasMatrix = pSceneInformationBuilder->GetProjectedAreasMatrix();
-    int numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
-    int numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
+    size_t numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
+    size_t numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
     mValues.resize( numberOfViewpoints );
     std::fill(mValues.begin(), mValues.end(), 0.0f);
     for( size_t currentViewpoint = 0; currentViewpoint < numberOfViewpoints; currentViewpoint++ )
     {
         unsigned int differentZero = 0;
-        for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+        for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
         {
             unsigned int a_z = projectedAreasMatrix->GetValue(currentViewpoint, currentPolygon);
             if( a_z != 0 )
