+++ src/viewpoint-measures/KullbackLeibler.cpp
@@ -15,20 +15,20 @@ KullbackLeibler::KullbackLeibler(const QString& pName): Measure(pName, false)
 void KullbackLeibler::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
 {
     const auto projectedAreasMatrix = pSceneInformationBuilder->GetProjectedAreasMatrix();
-    int numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
-    int numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
+    size_t numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
+    size_t numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
     mValues.resize( numberOfViewpoints );
     std::fill(mValues.begin(), mValues.end(), 0.0f);
     std::vector< float > serializedPolygonAreas = pSceneInformationBuilder->GetSerializedPolygonAreas();
     float sumAreaPolygons = 0.0f;
-    for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+    for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
     {
         sumAreaPolygons += serializedPolygonAreas.at(currentPolygon);
     }
     for( size_t currentViewpoint = 0; currentViewpoint < numberOfViewpoints; currentViewpoint++ )
     {
         unsigned int a_t = projectedAreasMatrix->GetSumPerViewpoint(currentViewpoint);
-        for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+        for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
         {
             unsigned int a_z = projectedAreasMatrix->GetValue(currentViewpoint, currentPolygon);
             if( a_z != 0 )
