+++ src/viewpoint-measures/VisibilityRatio.cpp
@@ -14,20 +14,20 @@ VisibilityRatio::VisibilityRatio(const QString& pName): Measure(pName, true)
 void VisibilityRatio::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
 {
     const auto projectedAreasMatrix = pSceneInformationBuilder->GetProjectedAreasMatrix();
-    int numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
-    int numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
+    size_t numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
+    size_t numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
     mValues.resize( numberOfViewpoints );
     std::fill(mValues.begin(), mValues.end(), 0.0f);
     std::vector< float > serializedPolygonAreas = pSceneInformationBuilder->GetSerializedPolygonAreas();
     float sumAreaPolygons = 0.0f;
-    for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+    for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
     {
         sumAreaPolygons += serializedPolygonAreas.at(currentPolygon);
     }
     for( size_t currentViewpoint = 0; currentViewpoint < numberOfViewpoints; currentViewpoint++ )
     {
         float sumAreasSeen = 0.0f;
-        for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+        for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
         {
             unsigned int a_z = projectedAreasMatrix->GetValue(currentViewpoint, currentPolygon);
             if( a_z != 0 )
